Bug fix on closing survey

+++ src/handlers/commands/stopSurvey.js
@@ -1,14 +1,18 @@
+const message = require('../../message')
+
 module.exports = async team => {
   const activeSurvey = await team.activeSurvey
   const scheduledSurvey = await team.scheduledSurvey
 
-  if (activeSurvey) {
-    await activeSurvey.close()
-    return 'Survey closed'
-  } else if (scheduledSurvey) {
-    await scheduledSurvey.delete()
-    return 'Survey unscheduled'
+  if (activeSurvey || scheduledSurvey) {
+    if (activeSurvey) {
+      await activeSurvey.close()
+    }
+    if (scheduledSurvey) {
+      await scheduledSurvey.delete()
+    }
+    return message.ok
   } else {
-    return 'No scheduled survey'
+    return 'No active or scheduled survey'
   }
 }
